+++ python-sdk/src/astro/databases/__init__.py
@@ -14,8 +14,8 @@ DEFAULT_CONN_TYPE_TO_MODULE_PATH = get_dict_with_module_names_to_dot_notations(P
 CUSTOM_CONN_TYPE_TO_MODULE_PATH = {
     "gcpbigquery": DEFAULT_CONN_TYPE_TO_MODULE_PATH["bigquery"],
     "google_cloud_platform": DEFAULT_CONN_TYPE_TO_MODULE_PATH["bigquery"],
-    "databricks": "astro.databricks.delta",
-    "delta": "astro.databricks.delta",
+    "databricks": "astro.databases.databricks.delta",
+    "delta": "astro.databases.databricks.delta",
 }
 CONN_TYPE_TO_MODULE_PATH = {
     **DEFAULT_CONN_TYPE_TO_MODULE_PATH,
