+++ python-sdk/src/astro/databases/databricks/load_file/jinja_templates/copy_into.py.jinja2
